test(core): consolidate duplicate check-secret and private-action-gate suites (#28811)
Consolidate parallel colocated and __tests__ suites for check-secret and
private-action-gate. Merge unique assertions: service unavailable,
per-key presence with missing list, string param handling, non-leakage,
isAutonomous true/false/type and metadata checks, private-allowed logic.
Delete redundant __tests__ files after verifying union preservation
(5+11 vs 2+3/6+5).

Closes #28811
